requirements.txt
# Backend & ML
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.3
sqlalchemy>=2.0.25
binance-connector>=3.5.0

# Machine Learning
torch>=2.2.0
scikit-learn>=1.4.0
matplotlib>=3.8.2
joblib>=1.3.2
pandas-ta>=0.3.14b0
xgboost>=2.0.3
optuna>=3.5.0

# Frontend
streamlit>=1.30.0
plotly>=5.18.0

# Testing
pytest>=8.0.0
httpx>=0.26.0
